Squashed commit of the following:
commit db7ce4ba9e9218150ad6100111347a88cd5e9e7a
Author: Oliver Tseng <otseng@yahoo.com>
Date:   Sat Aug 20 16:59:38 2022 -0400

    Fix tuples

commit 4309cc0409c911856817a154f91c2b54c2ccc681
Author: Oliver Tseng <otseng@yahoo.com>
Date:   Sat Aug 20 15:51:23 2022 -0400

    Update checkup.py

commit 4a7df3d82876ea71cf99a0c7aff1d115abf2fb26
Author: Oliver Tseng <otseng@yahoo.com>
Date:   Sat Aug 20 15:46:28 2022 -0400

    Update MainWindow.py

commit f43d43b02c5122918ed8b724d156814bf3101dd5
Author: Oliver Tseng <otseng@yahoo.com>
Date:   Sat Aug 20 09:35:47 2022 -0400

    Add pyinstaller

+++ util/ConfigUtil.py
@@ -20,6 +20,12 @@ class ConfigUtil:
         if not hasattr(config, "version") or current_version > config.version:
             config.version = current_version
 
+        # Determine if running in binary run mode or Python run mode
+        config.enableBinaryExecutionMode = False
+        if os.path.exists("enable_binary_execution_mode"):
+            config.enableBinaryExecutionMode = True
+            ConfigUtil.loadConfigIni()
+
         # Temporary configurations
         # Their values are not saved on exit.
         config.controlPanel = False
@@ -1365,10 +1371,6 @@ class ConfigUtil:
                 pat = file.readline().strip()
                 if pat:
                     config.githubAccessToken = pat
-                    logger = logging.getLogger('uba')
-                    logger.info(f"Using {patFile}")
-
-        ConfigUtil.loadColorConfig()
 
     # Save configurations on exit
     @staticmethod
@@ -1691,16 +1693,20 @@ class ConfigUtil:
             ("databaseConvertedOnStartup", config.databaseConvertedOnStartup),
             ("limitWorkspaceFilenameLength", config.limitWorkspaceFilenameLength),
         )
-        with open("config.py", "w", encoding="utf-8") as fileObj:
-            for name, value in configs:
-                fileObj.write("{0} = {1}\n".format(name, pprint.pformat(value)))
-            if hasattr(config, "translationLanguage"):
-                fileObj.write("{0} = {1}\n".format("translationLanguage", pprint.pformat(config.translationLanguage)))
-            if hasattr(config, "iModeSplitterSizes"):
-                fileObj.write("{0} = {1}\n".format("iModeSplitterSizes", pprint.pformat(config.iModeSplitterSizes)))
-            if hasattr(config, "pModeSplitterSizes"):
-                fileObj.write("{0} = {1}\n".format("pModeSplitterSizes", pprint.pformat(config.pModeSplitterSizes)))
-            # print("A copy of configurations is saved in file 'config.py'!")
+        if os.path.exists("enable_binary_execution_mode"):
+            with open("config.ini", "w", encoding="utf-8") as fileObj:
+                for name, value in configs:
+                    fileObj.write("{0} = {1}\n".format(name, str(value)))
+        else:
+            with open("config.py", "w", encoding="utf-8") as fileObj:
+                for name, value in configs:
+                    fileObj.write("{0} = {1}\n".format(name, pprint.pformat(value)))
+                if hasattr(config, "translationLanguage"):
+                    fileObj.write("{0} = {1}\n".format("translationLanguage", pprint.pformat(config.translationLanguage)))
+                if hasattr(config, "iModeSplitterSizes"):
+                    fileObj.write("{0} = {1}\n".format("iModeSplitterSizes", pprint.pformat(config.iModeSplitterSizes)))
+                if hasattr(config, "pModeSplitterSizes"):
+                    fileObj.write("{0} = {1}\n".format("pModeSplitterSizes", pprint.pformat(config.pModeSplitterSizes)))
 
     @staticmethod
     def getColorConfigFilename():
@@ -1715,3 +1721,20 @@ class ConfigUtil:
             with open(fileName, "r") as f:
                 settings = f.read()
                 exec(settings)
+    
+    def loadConfigIni():
+        if os.path.exists("config.ini"):
+            with open("config.ini", "r", encoding="utf-8") as fileObject:
+                for line in fileObject.readlines():
+                    parts = line.split("=", 1)
+                    key = parts[0].strip()
+                    value = parts[1].strip()
+                    valueType = getattr(config, key)
+                    if type(valueType) is str:
+                        setattr(config, key, str(value))
+                    elif type(valueType) is int:
+                        setattr(config, key, int(value))
+                    elif type(valueType) is float:
+                        setattr(config, key, float(value))
+                    else:
+                        setattr(config, key, eval(value))
